contextual-attachment: changed dataset conversion
